+++ sharktank/hip_kernels/kernels/topk_fp32_ukernel.c
@@ -58,7 +58,7 @@ extern "C" __global__ void topk_F32I32(const float *__restrict__ inputValues,
       int32_t curr_ind = warp_topk_indices[idx];
       bool swap = val > curr_val;
       float new_val = swap ? val : curr_val;
-      float new_ind = swap ? ind : curr_ind;
+      int32_t new_ind = swap ? ind : curr_ind;
       val = swap ? curr_val : val;
       ind = swap ? curr_ind : ind;
       warp_topk_vals[idx] = new_val;
@@ -84,7 +84,7 @@ extern "C" __global__ void topk_F32I32(const float *__restrict__ inputValues,
         int32_t curr_ind = warp_topk_indices[idx];
         bool swap = val > curr_val;
         float new_val = swap ? val : curr_val;
-        float new_ind = swap ? ind : curr_ind;
+        int32_t new_ind = swap ? ind : curr_ind;
         val = swap ? curr_val : val;
         ind = swap ? curr_ind : ind;
         warp_topk_vals[idx] = new_val;
@@ -109,7 +109,7 @@ extern "C" __global__ void topk_F32I32(const float *__restrict__ inputValues,
         int32_t curr_ind = warp_topk_indices[idx];
         bool swap = val > curr_val;
         float new_val = swap ? val : curr_val;
-        float new_ind = swap ? ind : curr_ind;
+        int32_t new_ind = swap ? ind : curr_ind;
         val = swap ? curr_val : val;
         ind = swap ? curr_ind : ind;
         warp_topk_vals[idx] = new_val;
